# Optimize hashing doubles by avoiding expensive NA checks