--- a/src/compare.c
+++ b/src/compare.c
@@ -44,29 +44,29 @@ int compare_scalar(SEXP x, R_len_t i, SEXP y, R_len_t j, bool na_equal) {
   case REALSXP: {
     double xi = REAL(x)[i], yj = REAL(y)[j];
     if (na_equal) {
-      enum vctrs_missingness x_missingness = dbl_missingness(xi);
-      enum vctrs_missingness y_missingness = dbl_missingness(yj);
+      enum vctrs_dbl_type x_type = dbl_type(xi);
+      enum vctrs_dbl_type y_type = dbl_type(yj);
 
-      switch (x_missingness) {
-      case vctrs_missingness_none: {
-        switch (y_missingness) {
-        case vctrs_missingness_none: return dcmp(xi, yj);
-        case vctrs_missingness_na: return 1L;
-        case vctrs_missingness_nan: return 1L;
+      switch (x_type) {
+      case vctrs_dbl_number: {
+        switch (y_type) {
+        case vctrs_dbl_number: return dcmp(xi, yj);
+        case vctrs_dbl_missing: return 1L;
+        case vctrs_dbl_nan: return 1L;
         }
       }
-      case vctrs_missingness_na: {
-        switch (y_missingness) {
-        case vctrs_missingness_none: return -1;
-        case vctrs_missingness_na: return 0;
-        case vctrs_missingness_nan: return 1;
+      case vctrs_dbl_missing: {
+        switch (y_type) {
+        case vctrs_dbl_number: return -1;
+        case vctrs_dbl_missing: return 0;
+        case vctrs_dbl_nan: return 1;
         }
       }
-      case vctrs_missingness_nan: {
-        switch (y_missingness) {
-        case vctrs_missingness_none: return -1;
-        case vctrs_missingness_na: return -1;
-        case vctrs_missingness_nan: return 0;
+      case vctrs_dbl_nan: {
+        switch (y_type) {
+        case vctrs_dbl_number: return -1;
+        case vctrs_dbl_missing: return -1;
+        case vctrs_dbl_nan: return 0;
         }
       }
       }