Today I learned how profile compiled C code using Apple's built in Instruments. I've got a few speed improvement PRs to submit based on my results. This is the first.

In `dbl_hash_scalar()` we call `R_IsNA()` and then `R_IsNaN()`, which is rather expensive. A more efficient approach is to use `isnan()` to first determine if a value is any of `NA` or `NaN`, and only call `R_IsNA()` to tell them apart after that.

This almost 2x's the hashing speed for doubles, which directly translates into improvements for the dictionary functions.

``` r
library(vctrs)
library(nycflights13)

vec_hash <- vctrs:::vec_hash

vec_unique_loc_old <- vctrs2::vec_unique_loc
vec_hash_old <- vctrs2:::vec_hash

# only double columns for the test
flights_dbl <- flights[, vapply(flights, is.double, logical(1))]
flights_dbl$time_hour <- NULL

bench::mark(
  vec_hash(flights_dbl),
  vec_hash_old(flights_dbl),
  iterations = 200
)
#> # A tibble: 2 x 6
#>   expression                     min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_hash(flights_dbl)       3.85ms   4.37ms      225.    1.29MB     6.95
#> 2 vec_hash_old(flights_dbl)   8.89ms   9.59ms      103.    1.29MB     3.18

bench::mark(
  vec_unique_loc(flights_dbl),
  vec_unique_loc_old(flights_dbl),
  iterations = 200
)
#> # A tibble: 2 x 6
#>   expression                         min median `itr/sec` mem_alloc
#>   <bch:expr>                      <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_unique_loc(flights_dbl)     34.5ms 39.1ms      24.8    8.56MB
#> 2 vec_unique_loc_old(flights_dbl)   41ms 47.1ms      20.9    8.55MB
#> # … with 1 more variable: `gc/sec` <dbl>
```

<sup>Created on 2019-11-05 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>

If you are curious, calling `vec_unique_loc()` looks like this with Instruments 

<img width="581" alt="Screen Shot 2019-11-05 at 3 24 14 PM" src="https://user-images.githubusercontent.com/19150088/68243387-64c68600-ffe0-11e9-9aa8-7eac02f61a0a.png">