@DavisVaughan 2019-11-06
Improvements from using `dbl_missing_indicator()` in `vec_equal()`:

``` r
library(vctrs)
x <- 1:1e6 + 0

# before
bench::mark(vec_equal(x, x, na_equal = TRUE))
#> # A tibble: 1 x 6
#>   expression                          min median `itr/sec` mem_alloc
#>   <bch:expr>                       <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_equal(x, x, na_equal = TRUE) 4.72ms 5.95ms      168.    3.84MB
#> # … with 1 more variable: `gc/sec` <dbl>

# after
bench::mark(vec_equal(x, x, na_equal = TRUE))
#> # A tibble: 1 x 6
#>   expression                          min median `itr/sec` mem_alloc
#>   <bch:expr>                       <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_equal(x, x, na_equal = TRUE) 2.07ms 2.61ms      353.    3.84MB
#> # … with 1 more variable: `gc/sec` <dbl>
```

Improvements from using `dbl_missing_indicator()` in `vec_compare()`:

``` r
library(vctrs)
x <- 1:1e6 + 0

# before
bench::mark(vec_compare(x, x, na_equal = TRUE))
#> # A tibble: 1 x 6
#>   expression                            min median `itr/sec` mem_alloc
#>   <bch:expr>                         <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_compare(x, x, na_equal = TRUE) 26.5ms 29.6ms      33.1    19.2MB
#> # … with 1 more variable: `gc/sec` <dbl>

# after
bench::mark(vec_compare(x, x, na_equal = TRUE))
#> # A tibble: 1 x 6
#>   expression                            min median `itr/sec` mem_alloc
#>   <bch:expr>                         <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_compare(x, x, na_equal = TRUE) 22.2ms 25.9ms      37.5    19.2MB
#> # … with 1 more variable: `gc/sec` <dbl>
```

`vec_compare()` should get a lot faster with some of the notes in #648